+++ docs/Whats-New.md
@@ -421,6 +421,7 @@ HideShakeEffects=false           ; boolean
 #### New:
 - Customized transport plane for teams (by FlyStar)
 - [Modify ammo on impact](New-or-Enhanced-Logics.md#modify-ammo-on-impact) (by FS-21)
+- [Passenger weapon count](New-or-Enhanced-Logics.md#passenger-weapon-count) (by weiyongxuan)
 
 #### Vanilla fixes:
 - Fixed the bug where a building with `Factory=BuildingType` owned by the AI did not play `ProductionAnim` when placing a produced building (by Noble_Fish)